backend/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
DateTime==4.3
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-ngrok==0.0.25
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pymongo==3.10.1
pytz==2019.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
zope.interface==5.1.0
requests==2.24.0
nltk==3.4.5
numpy==1.19.1
pyenchant==3.1.1
beautifulsoup4==4.9.0